fixed 2 Mac warnings

+++ Core/Source/BCKISSNCode.m
@@ -52,7 +52,7 @@
         }
         default:
         {
-            return [NSString stringWithFormat:@"%d", checkDigit];
+            return [NSString stringWithFormat:@"%d", (int)checkDigit];
             break;
         }
     }
